Store CLOBs and BLOBs in maps (StreamStore) to improve performance and avoid using a second connection. (WIP)

+++ trunk/h2/src/test/org/h2/test/db/TestLob.java
@@ -219,12 +219,14 @@ public class TestLob extends TestBase {
             prep.execute();
         }
         if (upgraded) {
-            if (config.memory) {
-                stat.execute("update information_schema.lob_map set pos=null");
-            } else {
-                stat.execute("alter table information_schema.lob_map drop column pos");
-                conn.close();
-                conn = getConnection("lob");
+            if (!config.mvStore) {
+                if (config.memory) {
+                    stat.execute("update information_schema.lob_map set pos=null");
+                } else {
+                    stat.execute("alter table information_schema.lob_map drop column pos");
+                    conn.close();
+                    conn = getConnection("lob");
+                }
             }
         }
         prep = conn.prepareStatement("select * from test where id = ?");
@@ -527,7 +529,7 @@ public class TestLob extends TestBase {
     }
 
     private void testDelete() throws Exception {
-        if (config.memory) {
+        if (config.memory || config.mvStore) {
             return;
         }
         deleteDb("lob");
